+++ src/main/java/org/ojalgo/matrix/decomposition/QR.java
@@ -23,6 +23,7 @@ package org.ojalgo.matrix.decomposition;
 
 import org.ojalgo.array.PlainArray;
 import org.ojalgo.matrix.store.MatrixStore;
+import org.ojalgo.matrix.store.PhysicalStore;
 import org.ojalgo.matrix.store.TransformableRegion;
 import org.ojalgo.scalar.ComplexNumber;
 import org.ojalgo.scalar.Quadruple;
@@ -95,6 +96,11 @@ public interface QR<N extends Comparable<N>> extends MatrixDecomposition.Updatab
         final MatrixStore<N> tmpStore = tmpQ.multiply(tmpR);
 
         return Access2D.equals(tmpStore, matrix, context);
+    }
+
+    @Override
+    default void ftran(final PhysicalStore<N> arg) {
+        this.getSolution(arg.copy(), arg);
     }
 
     MatrixStore<N> getQ();